+++ server/src/weaverbird/backends/pandas_executor/steps/utils/formula.py
@@ -15,6 +15,16 @@ _OP_MAP = {
 }
 
 
+class NullSeries(Series):
+    """A Series containing only null values.
+
+    This type allows us to know that a Series is filled with null values without having to check its content
+    """
+
+    def __init__(self, index: DataFrame.index):
+        Series.__init__(self, index=index, dtype="object")
+
+
 def _eval_operation(df: DataFrame, op: Operation) -> Series:
     return _OP_MAP[op.operator](_eval_expression(df, op.left), _eval_expression(df, op.right)).replace(
         [np.inf, -np.inf], np.nan
@@ -22,7 +32,9 @@ def _eval_operation(df: DataFrame, op: Operation) -> Series:
 
 
 def _eval_expression(df: DataFrame, expr: Expression) -> Series:
-    if isinstance(expr, Operation):
+    if expr is None:
+        return NullSeries(df.index)
+    elif isinstance(expr, Operation):
         return _eval_operation(df, expr)
     elif isinstance(expr, ColumnName):
         return df[expr.name]